Add option to define domains as whitelist instead of blacklist
Needs more checking and testing, but I think this is pretty much
there.

Adds a checkbox under the domains list text area that allows you
to specify the list as a whitelist as opposed to the default
blacklist.

Also re-worded some of the functions and variables from
"domainBlacklist" to "domainlist" to better reflect the new
functionality.

This will work better for me as there are only a few domains I
*need* to access when working, but a hell of a lot of others I
could procrastinate on.