Make whitelisting actually work
In my haste I'd not thought this through properly: Blacklisting
works my looping through the list of domains for each tab and if
it finds a match it blocks it. For whitelisting I was also looping
through the list of domains and only blocking if there wasn't a
match.

Of course this doesn't work as say you are visiting domainA and you
have whitelisted domainA and domainB. Well it will loop through, be
ok for domainA, but then check against the whitelisted domainB and
block. Doh!

So, this is messy code and there is probably a cleaner way to do it,
now set a whitelistok boolean to false. This is OR'd against each
domain in the whitelist and then only at the end of the loop if
whitelistok is still false does the blocking occur.