Automate Work / Break Cycle
Give the users a choice in preferences section (called 'auto mode')
which makes the extension to go in a work / break loop. This removes the
need to manually intervene to flip the work/break switch.

To Do: International strings. I have updated only locales\EN.

Future Improvements:
1) This option wont work well with Notifications. users won't see
notifications  because we
dismiss them on onStart when the next cycle begins. Users may
potentially notice
notification appear and disappear in a flash. This is minor annoyance or
sometimes un-noticeable. We can potentially uncheck Notifications, When
Auto-Mode is set.

2) I haven't addressed when the work / break loop expires. This will go
on in a loop. The user has to uncheck auto-mode option to break out of
the loop.